requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
soupsieve==2.6
spotipy==2.25.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
